Add files via upload
"CreateDB" does not need to be tested. It is just a simple way of quickly creating our databases.
"Database" is the program that allows querying of databases.
"ObjectInfo" is the object wrapper for object information.
"hsqldb" is the attached library.

All files are currently in the "HistoryPhone" package.
Note that they must be executed only together with the "hsqldb.jar".